requirements.txt
requests==2.31.0
websocket-client==1.6.4
pytest==7.4.3
pytest-asyncio==0.21.1
aiohttp==3.9.1
asyncio-throttle==1.0.2
pandas==2.1.4
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
colorama==0.4.6
tabulate==0.9.0
python-dotenv==1.0.0
pydantic==2.5.2
cryptography==41.0.8
